% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs (Autosaved).R
\name{distunlab}
\alias{distunlab}
\title{Get distance between two trees}
\usage{
distunlab(tree1, tree2)
}
\arguments{
\item{tree1}{A phylogenetic tree, phylo format or convertible}

\item{tree2}{A phylogenetic tree, phylo format or convertible}
}
\value{
labeldistance between the label sets of the two trees
}
\description{
Get distance between two trees
}
\examples{
distunlab(rtree(10), rtree(12))
}

