% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs.R
\name{getLabelFromPairs}
\alias{getLabelFromPairs}
\title{get label from the labels of the descendant nodes}
\usage{
getLabelFromPairs(twolabels, useAllCharacters = FALSE)
}
\arguments{
\item{useAllCharacters}{logical: use all character precision? default F}

\item{twolabels:}{vector of two labels}
}
\value{
a label, possibly hashed
}
\description{
get label from the labels of the descendant nodes
}
\examples{
getLabelFromPairs(c(2,3))
}

