% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs (Autosaved).R
\name{labeldistance}
\alias{labeldistance}
\title{Get distance between two vectors of labels}
\usage{
labeldistance(x, y)
}
\arguments{
\item{x}{A vector of tree labels (as created by treelabels)}

\item{y}{A vector of tree labels (as created by treelabels)}
}
\value{
D1 distance: the symmetric set difference between x and y
}
\description{
Get distance between two vectors of labels
}
\examples{
labeldistance(treelabels(rtree(10)), treelabels(rtree(12))
}

