% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs (Autosaved).R
\name{multiDistUnlab}
\alias{multiDistUnlab}
\title{All pairwise D1 distances for a list of trees}
\usage{
multiDistUnlab(trees, listoflabels = NA)
}
\arguments{
\item{trees}{A list of trees}

\item{listoflabels}{Alternatively, a list of labels. 
If listoflabels is NA (the default), this function
 first computes the labels for the set of trees. 
If listoflabels is not NA, the labels are compared directly and the trees are ignored.}
}
\value{
An object of class 'dist' containing tree-tree D1 distances
}
\description{
All pairwise D1 distances for a list of trees
}
\examples{
dd=multiDistUnlab(rmtree(10,12))
}

