% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NtoS.R
\name{nthPhylo}
\alias{nthPhylo}
\title{Obtain the tree for integer n}
\usage{
nthPhylo(n)
}
\arguments{
\item{n}{An integer}
}
\value{
phylo object tree corresponding to n
}
\description{
Obtain the tree for integer n
}
\examples{
nthPhylo(10)
}

