% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs (Autosaved).R
\name{plotlabels}
\alias{plotlabels}
\title{Plot a tree showing the labels}
\usage{
plotlabels(tree)
}
\arguments{
\item{tree}{A tree in phylo format
@return No variable is returned
@examples
plotlabels(rtree(10))}
}
\description{
Plot a tree showing the labels
}

