% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs.R
\name{prunetosize}
\alias{prunetosize}
\title{Prune a tree to the desired size, randomly uniformly dropping tips}
\usage{
prunetosize(tree, size)
}
\arguments{
\item{tree}{A tree in phylo format}

\item{size}{an integer, the size of the final tree}
}
\value{
a tree in phylo format, with the desired number of tips (size)
}
\description{
Prune a tree to the desired size, randomly uniformly dropping tips
}
\examples{
prunetosize(rtree(20),12)
}

