% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs.R
\name{treelabels}
\alias{treelabels}
\title{Create the labels for a tree}
\usage{
treelabels(tree, allChar = FALSE)
}
\arguments{
\item{tree}{A tree in phylo format (or convertible with as(tree,"phylo")}

\item{allChar}{Whether to use all character precision (default is false)}
}
\value{
A vector of labels for the tree including tips and internal nodes
}
\description{
Create the labels for a tree
}
\examples{
treelabels(rtree(10)
}

