% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs (Autosaved).R
\name{weightlabeldistance}
\alias{weightlabeldistance}
\title{Get weighted D2 distance between two vectors of labels}
\usage{
weightlabeldistance(lab1, lab2, weights = function(x) ifelse(is.numeric(x),
  1/(1 + log(x)), 1e-04))
}
\arguments{
\item{lab1}{A vector of tree labels (as created by treelabels)}

\item{lab2}{A vector of tree labels (as created by treelabels)}

\item{weights}{A function in the form ifelse(is.numeric(x),VALUE1,VALUE2) 
where VALUE1 is the weight applied to labels that have not been hashed 
(integers; fewer than 12 digits) and VALUE2 is the value if hashed. 
Default is (1/(1+log(x)) and 0.0001}
}
\value{
weighted D2 distance
}
\description{
Get weighted D2 distance between two vectors of labels
}
\examples{
weightlabeldistance(treelabels(rtree(10)), treelabels(rtree(12))
}

