% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabelledFuncs.R
\name{weightMultiDistUnlab}
\alias{weightMultiDistUnlab}
\title{All weighted pairwise distances for a list of trees}
\usage{
weightMultiDistUnlab(trees, listoflabels = NA, weights = function(x)
  ifelse(is.numeric(x), 1/(1 + log(x)), 1e-08))
}
\arguments{
\item{trees}{A list of trees}

\item{listoflabels}{Alternatively, a list of labels. If listoflabels is NA 
(the default), this function first computes the labels for the set of trees.
If listoflabels is not NA, the labels are compared directly and the trees are ignored.}

\item{weights}{A function in the form ifelse(is.numeric(x),VALUE1,VALUE2) 
where VALUE1 is the weight applied to labels that have not been hashed (fewer
than 12 digits) and VALUE2 is the value if hashed. Default is (1/(1+log(x)) and 1e-8}
}
\value{
An object of class 'dist' containing tree-tree distances
}
\description{
All weighted pairwise distances for a list of trees
}
\examples{
dd=weightmultiDistUnlab(rmtree(10,12))
}

